practice C language and test with memory
Today I learned a little bit about the C coding language. Things like
the '*' character being a pointer and the '&' character getting the
address. I learned that hardware is being virtualized so that the OS
can interact with it. The applications make system calls to the kernel
of the OS in order to do things like read/write, store information in
memory, etc. I learned about threads and how multiple threads can be
ran concurrently.